Switching to builder pattern, removing unnecessary test
I've switched the Dockerfile to use the "builder pattern" so that
the resulting docker image can be smaller (you aren't carrying
around the build dependencies, only the resulting executable).

Also "mkdir -p directory" does not need to be protected by
an "if [ ! -d directory ]", because it is a NOOP if the directory
already exists.